requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
bcrypt==4.1.2
pyjwt==2.8.0
openstacksdk==4.5.0